Added TrainerService and updated MemberService Business logic

+++ src/main/java/com/Gym/GymMembershipManagementSystem/dto/trainerDto/TrainerCreateDTO.java
@@ -3,6 +3,8 @@ package com.Gym.GymMembershipManagementSystem.dto.trainerDto;
 import com.Gym.GymMembershipManagementSystem.entity.Selections.GymSpecialization;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class TrainerCreateDTO {
 //    Add trainer
@@ -12,5 +14,5 @@ public class TrainerCreateDTO {
 
     private String phone;
 
-    private int salary;
+    private BigDecimal salary;
 }
